Fix argument loading in r2asm.lua
The previous version of r2asm would return "r2asm.lua:25: attempt to index local 'arg' (a number value)" when attempting to load a program
e.g. r2asm("quadratic.asm", 0xDEAD, "r2asm.dead.log")
This was because the argument loader would try to match numbers with regex. (in this case, the machine id: 0xDEAD)
This can be fixed by checking if the type of the argument, and if it's "number", then skipping the regex matching, and inserting the argument into the unnamed_args table